# Improve performance of long encoding