I've made a few changes to improve performance.
The benchmark was run on a Apple M3 Pro, the old and new methods where executed 1M times each with value N.

Here is the result of the benchmark:
| N | Old | New |
| - | - | - |
| 0 |260ms|237ms|
| -1 |248ms|227ms|
| 127 |480ms|357ms|
| 128 |481ms|355ms|
| 256 |481ms|355ms|
| -256 |485ms|356ms|
| 65536 |711ms|404ms|
| -65536 |710ms|397ms|
| 16777216 |946ms|442ms|
| -16777216 |947ms|447ms|
| 4294967296 |1184ms|474ms|
| -4294967296 |1185ms|474ms|
| 2305843009213693952 |2065ms|623ms|
| -2305843009213693952 |2088ms|626ms|
| 4611686018427387904 |2293ms|658ms|
| -4611686018427387904 |2060ms|624ms|
| 9223372036854775807 |2278ms|661ms|
| -9223372036854775808 |2299ms|663ms|